complete answer all layers module

+++ src/main/java/by/ghoncharko/selectioninterview/service/JpaHibernateQuestionService.java
@@ -9,6 +9,9 @@ import by.ghoncharko.selectioninterview.dto.QuestionDTOWithoutQuestionType;
 import by.ghoncharko.selectioninterview.dto.QuestionDtoForCreateOrUpdate;
 import by.ghoncharko.selectioninterview.entity.Question;
 import by.ghoncharko.selectioninterview.entity.QuestionType;
+import by.ghoncharko.selectioninterview.error.CannotCreateQuestionError;
+import by.ghoncharko.selectioninterview.error.CannotCreateQuestionTypeError;
+import by.ghoncharko.selectioninterview.error.CannotDeleteQuestionError;
 import by.ghoncharko.selectioninterview.error.CannotUpdateQuestionError;
 import by.ghoncharko.selectioninterview.mapper.QuestionMapper;
 import lombok.RequiredArgsConstructor;
@@ -26,10 +29,11 @@ import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
-public class JpaHibernateQuestionService implements QuestionService{
+public class JpaHibernateQuestionService implements QuestionService {
     private static final QuestionMapper questionMapper = QuestionMapper.INSTANCE;
     private final QuestionRepository questionRepository;
     private final QuestionTypeRepository questionTypeRepository;
+
     @Override
     @Transactional(readOnly = true)
     public Optional<QuestionDTOWithoutAnswersAndType> findByQuestionBodyIgnoreCase(String questionBody) {
@@ -75,8 +79,8 @@ public class JpaHibernateQuestionService implements QuestionService{
     @Override
     @Transactional(readOnly = true)
     public List<QuestionDTOWithoutAnswersAndType> findAllByDeletedIsFalse(Pageable pageable) {
-       Page<Question> questionPage = questionRepository.findAllByDeletedIsFalse(pageable);
-       return questionMapper.mapQuestionListToQuestionDtoWithoutAnswersAndTypeList(questionPage.getContent());
+        Page<Question> questionPage = questionRepository.findAllByDeletedIsFalse(pageable);
+        return questionMapper.mapQuestionListToQuestionDtoWithoutAnswersAndTypeList(questionPage.getContent());
     }
 
     @Override
@@ -117,12 +121,26 @@ public class JpaHibernateQuestionService implements QuestionService{
     @Override
     @Transactional
     public void deleteById(BigInteger id) {
+        Optional<Question> questionOptional = questionRepository.findById(id);
+        if (questionOptional.isEmpty()) {
+            throw new CannotDeleteQuestionError("Cannot delete question by id because question is not present in database questionId = " + id);
+        }
+        if (questionOptional.get().isDeleted()) {
+            throw new CannotDeleteQuestionError("Cannot delete question by id because question already is deleted questionId =" + id);
+        }
         questionRepository.deleteById(id);
     }
 
     @Override
     @Transactional
     public void deleteByQuestionBody(String body) {
+        Optional<Question> questionOptional = questionRepository.findByQuestionBody(body);
+        if (questionOptional.isEmpty()) {
+            throw new CannotDeleteQuestionError("Cannot delete question by question body because question is not present in database questionBody = " + body);
+        }
+        if (questionOptional.get().isDeleted()) {
+            throw new CannotDeleteQuestionError("Cannot delete question by question body because question already is deleted questionBody = " + body);
+        }
         questionRepository.deleteByQuestionBody(body);
     }
 
@@ -132,23 +150,28 @@ public class JpaHibernateQuestionService implements QuestionService{
         Question question = questionMapper.questionDtoForCreateOrUpdateToQuestion(questionDtoForCreateOrUpdate);
         question.setDateCreated(new Timestamp(new Date().getTime()));
         question.setLastDateUpdated(new Timestamp(new Date().getTime()));
-        Optional<QuestionType> questionTypeOptional = questionTypeRepository.findByQuestionTypeName(question.getQuestionType().getQuestionTypeName());
-        if(questionTypeOptional.isEmpty()){
-            QuestionType questionType = QuestionType.builder().
-                    questionTypeName(question.getQuestionType().getQuestionTypeName()).
-                    dateCreated(new Timestamp(new Date().getTime())).
-                    lastDateUpdated(new Timestamp(new Date().getTime())).
-                    deleted(false).
-                    build();
+        Optional<Question> questionOptional = questionRepository.findByQuestionBody(question.getQuestionBody());
+        if (questionOptional.isPresent()) {
+            throw new CannotCreateQuestionError("Cannot create question because question body is present questioBody = " + question.getQuestionBody());
+        }
+        Optional<QuestionType> questionTypeOptional = questionTypeRepository.findByQuestionTypeName(question.getQuestionType()
+                .getQuestionTypeName());
+        if (questionTypeOptional.isEmpty()) {
+            QuestionType questionType = QuestionType.builder()
+                    .questionTypeName(question.getQuestionType().getQuestionTypeName())
+                    .dateCreated(new Timestamp(new Date().getTime()))
+                    .lastDateUpdated(new Timestamp(new Date().getTime()))
+                    .deleted(question.getQuestionType().isDeleted())
+                    .build();
             QuestionType questionTypeAfterSave = questionTypeRepository.save(questionType);
             question.setQuestionType(questionTypeAfterSave);
-        }else {
-           QuestionType questionType = questionTypeOptional.get();
-           question.setQuestionType(questionType);
+        } else {
+            QuestionType questionType = questionTypeOptional.get();
+            question.setQuestionType(questionType);
         }
         Question questionAfterSave = questionRepository.save(question);
         QuestionDTO questionDTO = questionMapper.questionToQuestionDto(questionAfterSave);
-        if(questionDTO.getAnswers()==null){
+        if (questionDTO.getAnswers() == null) {
             questionDTO.setAnswers(new ArrayList<>());
         }
         return questionDTO;
@@ -159,23 +182,33 @@ public class JpaHibernateQuestionService implements QuestionService{
     public QuestionDTO update(QuestionDtoForCreateOrUpdate questionDtoForCreateOrUpdate) {
         Question question = questionMapper.questionDtoForCreateOrUpdateToQuestion(questionDtoForCreateOrUpdate);
         Optional<Question> questionFromDatabaseOptional = questionRepository.findById(question.getId());
-        if(questionFromDatabaseOptional.isPresent()){
+        if (questionFromDatabaseOptional.isPresent()) {
             Question questionFromDatabase = questionFromDatabaseOptional.get();
             question.setDateCreated(questionFromDatabase.getDateCreated());
             question.setLastDateUpdated(new Timestamp(new Date().getTime()));
-        }else {
-            throw new CannotUpdateQuestionError("Cannot update question is not present questionId= " + questionDtoForCreateOrUpdate.getId());
+        } else {
+            throw new CannotUpdateQuestionError("Cannot update question because is not present in database questionId= " + questionDtoForCreateOrUpdate.getId());
         }
-        Optional<QuestionType> questionTypeOptional = questionTypeRepository.findByQuestionTypeName(question.getQuestionType().getQuestionTypeName());
-        if(questionTypeOptional.isEmpty()){
-            throw new CannotUpdateQuestionError("Cannot update questionType is not present questionTypeId= " + questionDtoForCreateOrUpdate.getQuestionTypeDTO().getId());
+        Optional<Question> questionOptional = questionRepository.findByQuestionBody(question.getQuestionBody());
+        if (questionOptional.isPresent() && !questionOptional.get().getId().equals(question.getId())) {
+            throw new CannotUpdateQuestionError("Cannot update question because question body is present in database questonBody = " + question.getQuestionBody());
         }
-        QuestionType questionTypeFromOptional = questionTypeOptional.get();
-        questionTypeFromOptional.setLastDateUpdated(new Timestamp(new Date().getTime()));
-        question.setQuestionType(questionTypeFromOptional);
+        Optional<QuestionType> questionTypeOptional = questionTypeRepository.findByQuestionTypeName(question.getQuestionType()
+                .getQuestionTypeName());
+        if (questionTypeOptional.isPresent()) {
+            QuestionType questionTypeFromOptional = questionTypeOptional.get();
+            question.setQuestionType(questionTypeFromOptional);
+        } else {
+            QuestionType questionType = new QuestionType().builder()
+                    .questionTypeName(question.getQuestionType().getQuestionTypeName())
+                    .deleted(question.getQuestionType().isDeleted())
+                    .build();
+            QuestionType questionTypeAfterSave = questionTypeRepository.save(questionType);
+            question.setQuestionType(questionTypeAfterSave);
+        }
         Question questionAfterSave = questionRepository.save(question);
         QuestionDTO questionDTO = questionMapper.questionToQuestionDto(questionAfterSave);
-        if(questionDTO.getAnswers()==null){
+        if (questionDTO.getAnswers() == null) {
             questionDTO.setAnswers(new ArrayList<>());
         }
         return questionDTO;
